keySet(), Sets.intersection, isAvailable, protected validator & getters
