+++ server/src/test/java/org/elasticsearch/ingest/IngestSourceAndMetadataTests.java
@@ -65,42 +65,6 @@ public class IngestSourceAndMetadataTests extends ESTestCase {
         assertEquals(10000, md.getIfPrimaryTerm());
     }
 
-    public void testGetString() {
-        Map<String, Object> metadata = new HashMap<>();
-        metadata.put("a", "A");
-        metadata.put("b", new Object() {
-            @Override
-            public String toString() {
-                return "myToString()";
-            }
-        });
-        metadata.put("c", null);
-        metadata.put("d", 1234);
-        map = new IngestSourceAndMetadata(new HashMap<>(), new TestMetadata(metadata, allowAllValidators("a", "b", "c", "d")));
-        md = map.getMetadata();
-        assertNull(md.getString("c"));
-        assertNull(md.getString("no key"));
-        assertEquals("myToString()", md.getString("b"));
-        assertEquals("A", md.getString("a"));
-        assertEquals("1234", md.getString("d"));
-    }
-
-    public void testGetNumber() {
-        Map<String, Object> metadata = new HashMap<>();
-        metadata.put("a", Long.MAX_VALUE);
-        metadata.put("b", Double.MAX_VALUE);
-        metadata.put("c", "NaN");
-        metadata.put("d", null);
-        map = new IngestSourceAndMetadata(new HashMap<>(), new TestMetadata(metadata, allowAllValidators("a", "b", "c", "d")));
-        md = map.getMetadata();
-        assertEquals(Long.MAX_VALUE, md.getNumber("a"));
-        assertEquals(Double.MAX_VALUE, md.getNumber("b"));
-        IllegalArgumentException err = expectThrows(IllegalArgumentException.class, () -> md.getNumber("c"));
-        assertEquals("unexpected type for [c] with value [NaN], expected Number, got [java.lang.String]", err.getMessage());
-        assertNull(md.getNumber("d"));
-        assertNull(md.getNumber("no key"));
-    }
-
     public void testInvalidMetadata() {
         Map<String, Object> metadata = new HashMap<>();
         metadata.put("_version", Double.MAX_VALUE);
@@ -362,13 +326,5 @@ public class IngestSourceAndMetadataTests extends ESTestCase {
         public Object setValue(Object value) {
             throw new UnsupportedOperationException();
         }
-    }
-
-    private static Map<String, Metadata.Validator> allowAllValidators(String... keys) {
-        Map<String, Metadata.Validator> validators = new HashMap<>();
-        for (String key : keys) {
-            validators.put(key, (o, k, v) -> {});
-        }
-        return validators;
     }
 }
